DropDownEditor: Move 'contentTemplate' option to DropDownBox (T860163)
