+++ js/ui/drop_down_editor/ui.drop_down_editor.js
@@ -183,7 +183,6 @@ const DropDownEditor = TextBox.inherit({
             dropDownButtonTemplate: 'dropDownButton',
 
             fieldTemplate: null,
-            contentTemplate: null,
 
             /**
              * @name dxDropDownEditorOptions.openOnFieldClick
@@ -556,6 +555,8 @@ const DropDownEditor = TextBox.inherit({
         this._renderPopupContent();
     },
 
+    _renderPopupContent: noop,
+
     _renderPopup: function() {
         this._popup = this._createComponent(this._$popup, Popup, extend(this._popupConfig(), this._getInnerOptionsCache('dropDownOptions')));
 
@@ -603,7 +604,8 @@ const DropDownEditor = TextBox.inherit({
             showCloseButton: false,
             toolbarItems: this._getPopupToolbarItems(),
             onPositioned: this._popupPositionedHandler.bind(this),
-            fullScreen: false
+            fullScreen: false,
+            contentTemplate: null
         };
     },
 
@@ -653,27 +655,6 @@ const DropDownEditor = TextBox.inherit({
         }
 
         return this.callBase(positionRequest);
-    },
-
-    _renderPopupContent: function() {
-        const contentTemplate = this._getTemplateByOption('contentTemplate');
-
-        if(!(contentTemplate && this.option('contentTemplate'))) {
-            return;
-        }
-
-        const $popupContent = this._popup.$content();
-        const templateData = {
-            value: this._fieldRenderData(),
-            component: this
-        };
-
-        $popupContent.empty();
-
-        contentTemplate.render({
-            container: getPublicElement($popupContent),
-            model: templateData
-        });
     },
 
     _closeOutsideDropDownHandler: function({ target }) {
@@ -850,7 +831,6 @@ const DropDownEditor = TextBox.inherit({
                     this._invalidate();
                 }
                 break;
-            case 'contentTemplate':
             case 'acceptCustomValue':
             case 'openOnFieldClick':
                 this._invalidate();
