+++ testing/tests/DevExpress.ui.widgets.editors/dropDownEditor.tests.js
@@ -577,9 +577,6 @@ QUnit.module('focus policy', () => {
     QUnit.test('focusout to another editor should close current ddb (T832410)', function(assert) {
         const $dropDownEditor1 = $('#dropDownEditorLazy').dxDropDownEditor({
             items: [0, 1, 2],
-            contentTemplate() {
-                return $('<div>').attr('id', 'test-content');
-            },
             acceptCustomValue: true,
             focusStateEnabled: true,
             opened: true
@@ -939,24 +936,6 @@ QUnit.module('Templates', () => {
 
         assert.strictEqual($placeholder.length, 1, 'has only one placeholder');
         assert.strictEqual($placeholder.closest('.dx-textbox').length, 1, 'is textbox\'s placeholder');
-    });
-
-
-    QUnit.test('contentTemplate as render', function(assert) {
-        $('#dropDownEditorLazy').dxDropDownEditor({
-            contentTemplate(data, content) {
-                assert.equal(isRenderer(content), !!config().useJQuery, 'contentElement is correct');
-                $(content).addClass('drop-down-editor-content');
-                return $('<div>').text(data.component.option('value'));
-            },
-            value: 'test',
-            opened: true
-        });
-
-        const $dropDownContent = $('.drop-down-editor-content');
-
-        assert.equal($dropDownContent.length, 1, 'There is one dropDownEditor content element with custom class');
-        assert.equal($.trim($dropDownContent.text()), 'test', 'Correct content rendered');
     });
 
     QUnit.test('onValueChanged should be fired for each change by keyboard when fieldTemplate is used', function(assert) {
